requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile dev.in
#
annotated-types==0.6.0
    # via
    #   -r base.txt
    #   pydantic
anyio==3.7.1
    # via
    #   -r base.txt
    #   httpx
    #   openai
asgiref==3.7.2
    # via
    #   -r base.txt
    #   django
    #   django-cors-headers
asttokens==2.4.1
    # via
    #   -r base.txt
    #   stack-data
backcall==0.2.0
    # via
    #   -r base.txt
    #   ipython
beautifulsoup4==4.12.2
    # via
    #   -r base.txt
    #   django-bootstrap4
certifi==2023.11.17
    # via
    #   -r base.txt
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via
    #   -r base.txt
    #   cryptography
charset-normalizer==3.3.2
    # via
    #   -r base.txt
    #   requests
cryptography==41.0.5
    # via
    #   -r base.txt
    #   django-anymail
    #   pyjwt
decorator==5.1.1
    # via
    #   -r base.txt
    #   ipython
defusedxml==0.7.1
    # via
    #   -r base.txt
    #   python3-openid
distro==1.8.0
    # via
    #   -r base.txt
    #   openai
dj-rest-auth==5.0.2
    # via -r base.txt
django==3.2.12
    # via
    #   -r base.txt
    #   dj-rest-auth
    #   django-allauth
    #   django-anymail
    #   django-bootstrap4
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-filter
    #   django-phonenumber-field
    #   django-solo
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-admin-honeypot==1.1.0
    # via -r base.txt
django-allauth==0.58.2
    # via -r base.txt
django-anymail[sendgrid]==10.2
    # via -r base.txt
django-bootstrap4==23.2
    # via -r base.txt
django-cors-headers==4.3.1
    # via -r base.txt
django-debug-toolbar==4.2.0
    # via -r base.txt
django-environ==0.11.2
    # via -r base.txt
django-filter==23.3
    # via -r base.txt
django-phonenumber-field==7.2.0
    # via -r base.txt
django-phonenumbers==1.0.1
    # via -r base.txt
django-solo==2.1.0
    # via -r base.txt
djangorestframework==3.14.0
    # via
    #   -r base.txt
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-camel-case==1.4.2
    # via -r base.txt
djangorestframework-simplejwt==4.5.0
    # via -r base.txt
drf-yasg==1.21.7
    # via -r base.txt
exceptiongroup==1.1.3
    # via
    #   -r base.txt
    #   anyio
executing==2.0.1
    # via
    #   -r base.txt
    #   stack-data
gitdb==4.0.11
    # via
    #   -r base.txt
    #   gitpython
gitpython==3.1.40
    # via -r base.txt
gunicorn==21.2.0
    # via -r dev.in
h11==0.14.0
    # via
    #   -r base.txt
    #   httpcore
httpcore==1.0.2
    # via
    #   -r base.txt
    #   httpx
httpx==0.25.1
    # via
    #   -r base.txt
    #   openai
idna==3.4
    # via
    #   -r base.txt
    #   anyio
    #   httpx
    #   requests
importlib-metadata==6.8.0
    # via -r base.txt
inflection==0.5.1
    # via
    #   -r base.txt
    #   drf-yasg
ipython==8.12.3
    # via -r base.txt
jedi==0.19.1
    # via
    #   -r base.txt
    #   ipython
matplotlib-inline==0.1.6
    # via
    #   -r base.txt
    #   ipython
oauthlib==3.2.2
    # via
    #   -r base.txt
    #   requests-oauthlib
openai==1.3.3
    # via -r base.txt
packaging==23.2
    # via
    #   -r base.txt
    #   drf-yasg
    #   gunicorn
parso==0.8.3
    # via
    #   -r base.txt
    #   jedi
pexpect==4.8.0
    # via
    #   -r base.txt
    #   ipython
phonenumbers==8.13.25
    # via
    #   -r base.txt
    #   django-phonenumbers
pickleshare==0.7.5
    # via
    #   -r base.txt
    #   ipython
pillow==10.1.0
    # via -r base.txt
prompt-toolkit==3.0.41
    # via
    #   -r base.txt
    #   ipython
psycopg2-binary==2.9.9
    # via -r dev.in
ptyprocess==0.7.0
    # via
    #   -r base.txt
    #   pexpect
pure-eval==0.2.2
    # via
    #   -r base.txt
    #   stack-data
pycparser==2.21
    # via
    #   -r base.txt
    #   cffi
pydantic==2.5.1
    # via
    #   -r base.txt
    #   openai
pydantic-core==2.14.3
    # via
    #   -r base.txt
    #   pydantic
pygments==2.17.0
    # via
    #   -r base.txt
    #   ipython
pyjwt[crypto]==1.7.0
    # via
    #   -r base.txt
    #   django-allauth
    #   djangorestframework-simplejwt
python3-openid==3.2.0
    # via
    #   -r base.txt
    #   django-allauth
pytz==2023.3.post1
    # via
    #   -r base.txt
    #   django
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0.1
    # via
    #   -r base.txt
    #   drf-yasg
regex==2023.10.3
    # via
    #   -r base.txt
    #   tiktoken
requests==2.31.0
    # via
    #   -r base.txt
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
    #   tiktoken
requests-oauthlib==1.3.1
    # via
    #   -r base.txt
    #   django-allauth
six==1.16.0
    # via
    #   -r base.txt
    #   asttokens
smmap==5.0.1
    # via
    #   -r base.txt
    #   gitdb
sniffio==1.3.0
    # via
    #   -r base.txt
    #   anyio
    #   httpx
soupsieve==2.5
    # via
    #   -r base.txt
    #   beautifulsoup4
sqlparse==0.4.4
    # via
    #   -r base.txt
    #   django
    #   django-debug-toolbar
stack-data==0.6.3
    # via
    #   -r base.txt
    #   ipython
tiktoken==0.5.1
    # via -r base.txt
tqdm==4.66.1
    # via
    #   -r base.txt
    #   openai
traitlets==5.13.0
    # via
    #   -r base.txt
    #   ipython
    #   matplotlib-inline
typing-extensions==4.8.0
    # via
    #   -r base.txt
    #   annotated-types
    #   asgiref
    #   ipython
    #   openai
    #   pydantic
    #   pydantic-core
uritemplate==4.1.1
    # via
    #   -r base.txt
    #   drf-yasg
urllib3==2.1.0
    # via
    #   -r base.txt
    #   django-anymail
    #   requests
wcwidth==0.2.10
    # via
    #   -r base.txt
    #   prompt-toolkit
wheel==0.41.3
    # via -r base.txt
zipp==3.17.0
    # via
    #   -r base.txt
    #   importlib-metadata
